- Move parser fixtures to their own conftest.py
